enhance close budget pie chart
