+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tidebit-defi",
-  "version": "0.9.0+82",
+  "version": "0.9.0+83",
   "private": true,
   "type": "module",
   "scripts": {
