+++ src/constants/display.ts
@@ -29,3 +29,16 @@ export const MONTH_SHORT_NAME = [
 ];
 
 export const DEFAULT_PIC_URL = '/elements/default_pic.png';
+
+// Info: (20260116 - Julian) 定義圓餅圖顏色變數
+export const PIE_CHART_LABEL_COLOR_PROPERTY = '--color-text-neutral-primary';
+export const PIE_CHART_FILL_COLORS_PROPERTIES = [
+  '--color-surface-support-strong-maple',
+  '--color-surface-support-strong-taro',
+  '--color-surface-support-strong-rose',
+  '--color-surface-support-strong-green',
+  '--color-surface-support-strong-indigo',
+  '--color-surface-support-strong-pink',
+  '--color-surface-support-strong-baby',
+];
+export const PIE_CHART_EMPTY_COLOR_PROPERTY = '--color-loading-indicator-surface-base';
